% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ksic.R
\name{is_ksic}
\alias{is_ksic}
\title{Check for Valid KSIC Codes}
\usage{
is_ksic(ksic)
}
\arguments{
\item{ksic}{character. A vector of KSIC codes to check.}
}
\value{
A data.frame with the input codes and logical columns (C9, C10, C11) indicating validity.
}
\description{
Checks if the input codes are valid for the given KSIC revision (9th, 10th, or 11th).
}
