% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSIC-package.R
\docType{package}
\name{KSIC-package}
\alias{KSIC}
\alias{KSIC-package}
\title{KSIC: Korea Standard Industry Code (KSIC) Data}
\description{
This package offers a comprehensive toolkit for working with the Korea Standard Industrial Classification (KSIC). It includes complete datasets for the 9th, 10th, and 11th revisions of the KSIC. Key features include functions to search codes and names by keyword, convert codes across different revisions, validate KSIC codes, and navigate the classification hierarchy by identifying parent or child categories. The package is designed to support statistical analysis, data processing, and research related to South Korea's industrial classification standards.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/urbanjj/KSIC}
  \item Report bugs at \url{https://github.com/urbanjj/KSIC/issues}
}

}
\author{
\strong{Maintainer}: Jongjin Yun \email{jongjin@uos.ac.kr}

}
\keyword{internal}
