% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksic_10_to_11.R
\docType{data}
\name{ksic_10_to_11}
\alias{ksic_10_to_11}
\title{KSIC 10th to 11th Concordance Table}
\format{
A data frame with columns for 10th codes, 11th codes, and connection details.
}
\usage{
ksic_10_to_11
}
\description{
A concordance table for converting 10th revision KSIC codes to 11th revision KSIC codes.
}
\examples{
  ksic_10_to_11
}
\keyword{datasets}
