% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksic_10_to_9.R
\docType{data}
\name{ksic_10_to_9}
\alias{ksic_10_to_9}
\title{KSIC 10th to 9th Concordance Table}
\format{
A data frame with 6 columns:
\describe{
\item{ksic10_cd}{10th revision KSIC code.}
\item{ksic10_nm}{10th revision KSIC name.}
\item{ksic9_cd}{9th revision KSIC code.}
\item{ksic9_nm}{9th revision KSIC name.}
\item{con}{Type of connection.}
\item{detail}{Additional details about the connection.}
}
}
\usage{
ksic_10_to_9
}
\description{
A concordance table for converting 10th revision KSIC codes to 9th revision KSIC codes.
}
\examples{
  ksic_10_to_9
}
\keyword{datasets}
