% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksic_11_to_10.R
\docType{data}
\name{ksic_11_to_10}
\alias{ksic_11_to_10}
\title{KSIC 11차-10차 연계표 / KSIC 11th to 10th Concordance Table}
\format{
A data frame with columns for 11th codes, 10th codes, and connection details.
}
\source{
\url{https://kssc.kostat.go.kr}
}
\usage{
ksic_11_to_10
}
\description{
11차 KSIC 코드를 10차 KSIC 코드로 변환하기 위한 연계표입니다.
A concordance table for converting 11th revision KSIC codes to 10th revision KSIC codes.
}
\examples{
  ksic_11_to_10
}
\keyword{datasets}
