% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksic_9_to_10.R
\docType{data}
\name{ksic_9_to_10}
\alias{ksic_9_to_10}
\title{Korea Standard Industry Code Connection 9th to 10th}
\format{
Data frame with columns
\describe{
\item{ksic9_cd}{from code.}
\item{ksic9_nm}{from name.}
\item{ksic10_cd}{to code.}
\item{ksic10_nm}{to name.}
\item{con}{type of connection.}
\item{detail}{detail.}
}
}
\source{
\url{https://kssc.kostat.go.kr}
}
\usage{
ksic_9_to_10
}
\description{
Korea Standard Industry Code Connection 9th to 10th
}
\examples{
  ksic_9_to_10
}
\keyword{datasets}
