% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksic_find.R
\name{ksic_find}
\alias{ksic_find}
\title{Find KSIC Information by Code}
\usage{
ksic_find(codes)
}
\arguments{
\item{codes}{character. A vector of KSIC codes.}
}
\value{
A data.frame of matching KSIC codes and names.
}
\description{
Searches for KS.IC information by code.
}
