% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksicTreeDB.R
\docType{data}
\name{ksicTreeDB}
\alias{ksicTreeDB}
\title{Korea Standard Industry Code Tree Data}
\format{
A data frame with columns for each classification level (1 to 5 digits) and the KSIC revision.
}
\usage{
ksicTreeDB
}
\description{
A dataset representing the hierarchical structure of the 9th and 10th KSIC. It includes parent codes for each 5-digit classification.
}
\examples{
  ksicTreeDB
}
\keyword{datasets}
