% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSIC-package.R
\docType{package}
\name{KSIC-package}
\alias{KSIC}
\alias{KSIC-package}
\title{KSIC: Korea Standard Industrial Classification (KSIC)}
\description{
Provides tools for working with the Korea Standard Industrial Classification (KSIC). Includes datasets for the 9th, 10th, and 11th revisions. Functions include searching codes and names by keyword, converting codes across revisions, validating KSIC codes, and navigating the classification hierarchy (e.g., identifying parent or child categories). Intended for use in statistical analysis, data processing, and research involving South Korea’s industrial classification system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/urbanjj/KSIC}
  \item Report bugs at \url{https://github.com/urbanjj/KSIC/issues}
}

}
\author{
\strong{Maintainer}: Jongjin Yun \email{jongjin@uos.ac.kr} (\href{https://orcid.org/0009-0000-0161-2285}{ORCID})

}
\keyword{internal}
