% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSIC-package.R
\docType{package}
\name{KSIC-package}
\alias{KSIC}
\alias{KSIC-package}
\title{KSIC: Korea Standard Industry Code Data}
\description{
Korea Standard Industry Code Data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jjyunnn/KSIC}
  \item Report bugs at \url{https://github.com/jjyunnn/KSIC/issues}
}

}
\author{
\strong{Maintainer}: Jongjin Yun \email{jjjyunn@gmail.ac.kr} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
