% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSIC-package.R
\docType{package}
\name{KSIC-package}
\alias{KSIC}
\alias{KSIC-package}
\title{KSIC: Korea Standard Industry Code (KSIC) Data}
\description{
Provides data and functions for using the Korea Standard Industrial Classification (KSIC).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/urbanjj/KSIC}
  \item Report bugs at \url{https://github.com/urbanjj/KSIC/issues}
}

}
\author{
\strong{Maintainer}: Jongjin Yun \email{jongjin@uos.ac.kr}

}
\keyword{internal}
