% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksic_10_to_9.R
\docType{data}
\name{ksic_10_to_9}
\alias{ksic_10_to_9}
\title{KSIC 10th to 9th Concordance Table}
\format{
A data frame with columns for 10th codes, 9th codes, and connection details.
}
\usage{
ksic_10_to_9
}
\description{
A concordance table for converting 10th revision KSIC codes to 9th revision KSIC codes.
}
\examples{
  ksic_10_to_9
}
\keyword{datasets}
