% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksic_10_to_9.R
\docType{data}
\name{ksic_10_to_9}
\alias{ksic_10_to_9}
\title{Korea Standard Industry Code Connection 10th to 9th}
\format{
Data frame with columns
\describe{
\item{ksic10_cd}{from code.}
\item{ksic10_nm}{from name.}
\item{ksic9_cd}{to code.}
\item{ksic9_nm}{to name.}
\item{con}{type of connection.}
\item{detail}{detail.}
}
}
\source{
\url{https://kssc.kostat.go.kr}
}
\usage{
ksic_10_to_9
}
\description{
Korea Standard Industry Code Connection 10th to 9th
}
\examples{
  ksic_10_to_9
}
\keyword{datasets}
