% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksic_10_to_9.R
\docType{data}
\name{ksic_10_to_9}
\alias{ksic_10_to_9}
\title{KSIC 10차-9차 연계표 / KSIC 10th to 9th Concordance Table}
\format{
A data frame with columns for 10th codes, 9th codes, and connection details.
}
\source{
\url{https://kssc.kostat.go.kr}
}
\usage{
ksic_10_to_9
}
\description{
10차 KSIC 코드를 9차 KSIC 코드로 변환하기 위한 연계표입니다.
A concordance table for converting 10th revision KSIC codes to 9th revision KSIC codes.
}
\examples{
  ksic_10_to_9
}
\keyword{datasets}
