% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksic_11_to_10.R
\docType{data}
\name{ksic_11_to_10}
\alias{ksic_11_to_10}
\title{KSIC 11th to 10th Concordance Table}
\format{
A data frame with columns for 11th codes, 10th codes, and connection details.
}
\usage{
ksic_11_to_10
}
\description{
A concordance table for converting 11th revision KSIC codes to 10th revision KSIC codes.
}
\examples{
  ksic_11_to_10
}
\keyword{datasets}
