% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksicDB.R
\docType{data}
\name{ksicDB}
\alias{ksicDB}
\title{Korea Standard Industry Code Data}
\format{
Data frame with columns
\describe{
\item{cd}{code.}
\item{nm}{name.}
\item{digit}{digit=1,2,3,4,5.}
\item{ksic_C}{9 or 10.}
}
}
\source{
\url{https://kssc.kostat.go.kr}
}
\usage{
ksicDB
}
\description{
Korea Standard Industry Code Data
}
\examples{
  ksicDB
}
\keyword{datasets}
