% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksicDB.R
\docType{data}
\name{ksicDB}
\alias{ksicDB}
\title{Korea Standard Industry Code Data}
\format{
A data frame with columns:
\describe{
\item{cd}{Classification code.}
\item{nm}{Classification name.}
\item{eng_nm}{English classification name.}
\item{digit}{Digit of the classification (1-5).}
\item{ksic_C}{KSIC revision (C9 or C10).}
}
}
\source{
\url{https://kssc.kostat.go.kr}
}
\usage{
ksicDB
}
\description{
A dataset containing the codes, names, revisions, and digits of the 9th and 10th Korea Standard Industrial Classification (KSIC).
}
\examples{
  ksicDB
}
\keyword{datasets}
