% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksicTreeDB.R
\docType{data}
\name{ksicTreeDB}
\alias{ksicTreeDB}
\title{Korea Standard Industry Code Data Tree Structure}
\format{
Data frame with columns
\describe{
\item{ksic1_cd}{digit1 code.}
\item{ksic1_nm}{digit1 name.}
\item{ksic2_cd}{digit2 code.}
\item{ksic2_nm}{digit2 name.}
\item{ksic3_cd}{digit3 code.}
\item{ksic3_nm}{digit3 name.}
\item{ksic4_cd}{digit4 code.}
\item{ksic4_nm}{digit4 name.}
\item{ksic5_cd}{digit5 code.}
\item{ksic5_nm}{digit5 name.}
\item{ksic_C}{9 or 10.}
}
}
\source{
\url{https://kssc.kostat.go.kr}
}
\usage{
ksicTreeDB
}
\description{
Korea Standard Industry Code Data
}
\examples{
  ksicTreeDB
}
\keyword{datasets}
